+++ src/pages/map/pixelmeasures.vue
